Exploring Object Destructoring
Exploring an ever so usful es6 syntax that makes object referencing quicker & much eaiser. With it we can set a default value if the name does not exist to avoid NaN. Also we can rename the object name to a new name as if it was a variable all on one line. This is very handy & I will be taking full advantage of this syntax in the very near future. & Fuck the 99 center store.